[presenters-dispatcher] improv: rename package
