+++ moviper/src/main/java/com/mateuszkoslacz/moviper/presentersdispatcher/MoviperPresentersDispatcher.java
@@ -1,4 +1,4 @@
-package com.mateuszkoslacz.moviper.server;
+package com.mateuszkoslacz.moviper.presentersdispatcher;
 
 import android.app.Activity;
 import android.app.Fragment;
