+++ sample-super-rx-ai-kotlin/src/main/java/com/mateuszkoslacz/moviper/rxsample/viper/view/activity/UserDetailsActivity.kt
@@ -10,7 +10,7 @@ import com.mateuszkoslacz.moviper.iface.presenter.ViperPresenter
 import com.mateuszkoslacz.moviper.rxsample.R
 import com.mateuszkoslacz.moviper.rxsample.viper.contract.UserDetailsContract
 import com.mateuszkoslacz.moviper.rxsample.viper.entity.User
-import com.mateuszkoslacz.moviper.server.MoviperPresentersDispatcher
+import com.mateuszkoslacz.moviper.presentersdispatcher.MoviperPresentersDispatcher
 import kotlinx.android.synthetic.main.activity_user_details.*
 import rx.Observable
 import rx.subjects.PublishSubject
