+++ sample-super-rx-ai/src/test/java/com/mateuszkoslacz/moviper/rxsample/viper/routing/ListingRoutingRobolectricTest.java
@@ -7,8 +7,8 @@ import com.mateuszkoslacz.moviper.rxsample.viper.entity.User;
 import com.mateuszkoslacz.moviper.rxsample.viper.presenter.UserDetailsPresenter;
 import com.mateuszkoslacz.moviper.rxsample.viper.view.activity.ListingActivity;
 import com.mateuszkoslacz.moviper.rxsample.viper.view.activity.UserDetailsActivity;
-import com.mateuszkoslacz.moviper.server.ActivityStarter;
-import com.mateuszkoslacz.moviper.server.MoviperPresentersDispatcher;
+import com.mateuszkoslacz.moviper.presentersdispatcher.ActivityStarter;
+import com.mateuszkoslacz.moviper.presentersdispatcher.MoviperPresentersDispatcher;
 
 import org.junit.Before;
 import org.junit.Test;
