+++ sample-super-rx-databinding/src/main/java/com/mateuszkoslacz/moviper/rxsample/viper/view/activity/UserDetailsActivity.java
@@ -13,7 +13,7 @@ import com.mateuszkoslacz.moviper.rxsample.R;
 import com.mateuszkoslacz.moviper.rxsample.databinding.ActivityUserDetailsBinding;
 import com.mateuszkoslacz.moviper.rxsample.viper.contract.UserDetailsContract;
 import com.mateuszkoslacz.moviper.rxsample.viper.entity.User;
-import com.mateuszkoslacz.moviper.server.MoviperPresentersDispatcher;
+import com.mateuszkoslacz.moviper.presentersdispatcher.MoviperPresentersDispatcher;
 
 import rx.Observable;
 import rx.subjects.PublishSubject;
